reorganize test classes into transaction package
